addressing last of code review comments

+++ test/edu/washington/escience/myria/operator/apply/ApplyNgramTest.java
@@ -48,7 +48,7 @@ public class ApplyNgramTest {
 
     Apply apply = new Apply(new BatchTupleSource(input), expressions.build());
     apply.open(TestEnvVars.get());
-    long rowIdx = 0;
+    int rowIdx = 0;
     while (!apply.eos()) {
       TupleBatch result = apply.nextReady();
       if (result != null) {
@@ -87,7 +87,7 @@ public class ApplyNgramTest {
 
     Apply apply = new Apply(new BatchTupleSource(input), expressions.build(), true);
     apply.open(TestEnvVars.get());
-    long rowIdx = 0;
+    int rowIdx = 0;
     while (!apply.eos()) {
       TupleBatch result = apply.nextReady();
       if (result != null) {
@@ -96,7 +96,7 @@ public class ApplyNgramTest {
         for (int batchIdx = 0; batchIdx < result.numTuples(); ++batchIdx, ++rowIdx) {
           char[] ngramChars = new char[] {(char) rowIdx, (char) (rowIdx + 1), (char) (rowIdx + 2)};
           String ngram = new String(ngramChars);
-          int fltmapid = (int) rowIdx;
+          int fltmapid = rowIdx;
           assertEquals(ngram, result.getString(0, batchIdx));
           assertEquals(fltmapid, result.getInt(1, batchIdx));
         }
